authors: blueprint fix
* Reorganizes blueprints to make user facing forms work again.
